arreglo errores. compila

+++ src/algorithms/poisson/pois_factorialFormula.cpp
@@ -13,7 +13,7 @@ namespace algorithms::poisson::factorialFormula {
 
     float cdf(float lambda, unsigned int k){
         int res = 0;
-        for(int i = 0; i <= k; i++){
+        for(size_t i = 0; i <= k; i++){
             res += pmf(lambda, i);
         }
         return res;
